apps/backend/requirements.txt
Flask>=2.0.0
flask_cors>=3.0.0
SQLAlchemy>=1.4.0 
PyYAML~=6.0.2
werkzeug~=3.0.1
Pillow~=10.1.0
flask-restx>=1.3.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
flask_sqlalchemy
gunicorn>=21.0.0
requests>=2.31.0
email_validator>=1.3.0
redis>=4.5.0
flask-session>=0.5.0